refactor(cli): make jan agent remote-only and free of Tauri/GTK dependencies
The jan CLI could start local models (llama-server router) and, because
`tauri` was an unconditional dependency, linked the whole GUI stack: gtk
and gdk enter the graph through tauri itself via muda, so merely not
calling into Tauri was not enough to keep them out.

Remove local inference from the CLI and make the Tauri stack optional.

Remote-only:
- Drop the router lifecycle (ensure_router_and_load, discover_llamacpp_binary,
  spawn_local_router_if_needed), the router.preset.ini generator, and local
  model.yml discovery. `jan cli models list` now prints the models exposed by
  configured providers instead of installed GGUF/MLX files.
- Gate OrchestrationArgs::{llama_state, mlx_sessions} and the MLX/router
  branches of resolve_upstream_for_model out of the cli build. Desktop local
  inference is unchanged.
- Local models remain usable through the desktop app's API server, which is an
  ordinary remote provider here. Reachability is "has a non-empty base_url"
  (providers::is_cli_reachable), so base_url-less engine descriptors are
  filtered out of /model, `models list`, and `agent status`, and
  unreachable_local_provider rejects them up front with a pointer to the
  API-server setup.
- Fix a nondeterministic provider lookup: when several providers list the same
  model id (a Jan API server plus the local engine descriptor), HashMap
  iteration order decided the winner, so a reachable upstream was picked at
  random. The CLI now prefers a provider with a usable base_url.

Tauri-free:
- `tauri`, every tauri-plugin-*, tauri-build, rfd, gtk, zbus and fix-path-env
  become optional deps behind a new `tauri-app` feature, enabled by desktop,
  test-tauri and mobile. Dependency count drops 755 -> 458 and no
  gtk/gdk/muda/wry/webkit2gtk remains in the CLI graph.
- Desktop-only modules are #[cfg(not(feature = "cli"))]: core::{server,
  downloads, filesystem, system, updater}, the commands.rs IPC layers,
  AppState, mcp::{helpers, lockfile}, threads::commands, app::settings_store.
- Shared tauri-free logic is extracted rather than duplicated: the new
  core::openai_schema holds the tool-schema normalization and key-rotation
  helpers that lived in proxy.rs, and the MCP config parsers move from
  mcp::helpers to mcp::models (with their tests, which now run in both
  configs).
- tauri-plugin-{websearch,vector-db} expose their cores (provider, db) with
  `tauri` as a default-on optional feature, so the agent's web tools and memory
  work without it. arboard stays: its wayland/x11 crates are pure-Rust protocol
  clients, not GUI toolkits.

Build system:
- `cli` and `tauri-app` are mutually exclusive; a compile_error! says so,
  because `--features cli` alone still implies `default`. The Makefile, the
  three CLI workflows and build-tui.sh now pass --no-default-features.
- CI asserts the CLI graph stays free of tauri/gtk/gdk/muda/wry/webkit2gtk, and
  runs the cli-config test suite, which nothing exercised before.

Verified: clippy -D warnings and the test suites pass in both configs
(447 cli, 548 desktop); jan-desktop still builds; end-to-end agent runs with
tool calls against remote providers.
