+++ .github/workflows/jan-linter-and-test.yml
@@ -109,6 +109,9 @@ jobs:
         run: |
           cargo llvm-cov clean --workspace --manifest-path src-tauri/Cargo.toml
           cargo llvm-cov --no-report --manifest-path src-tauri/Cargo.toml --no-default-features --features test-tauri -- --test-threads=1
+          # The headless CLI is a separate, mutually-exclusive feature config
+          # (no Tauri); its modules are compiled out of the run above.
+          cargo llvm-cov --no-report --manifest-path src-tauri/Cargo.toml --no-default-features --features cli --lib -- --test-threads=1
           cargo llvm-cov --no-report --manifest-path src-tauri/plugins/tauri-plugin-hardware/Cargo.toml
           cargo llvm-cov --no-report --manifest-path src-tauri/plugins/tauri-plugin-llamacpp/Cargo.toml
           cargo llvm-cov --no-report --manifest-path src-tauri/utils/Cargo.toml
