+++ src-tauri/plugins/tauri-plugin-vector-db/Cargo.toml
@@ -10,8 +10,14 @@ rust-version = "1.77.2"
 exclude = ["/examples", "/dist-js", "/guest-js", "/node_modules"]
 links = "tauri-plugin-vector-db"
 
+[features]
+default = ["tauri"]
+# The Tauri command/plugin surface. Disabled by the headless `jan` CLI, which
+# consumes `db` directly and must not link any GUI crate.
+tauri = ["dep:tauri"]
+
 [dependencies]
-tauri = { version = "2.8.5", default-features = false }
+tauri = { version = "2.8.5", default-features = false, optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "2.0"
