+++ src-tauri/plugins/tauri-plugin-websearch/Cargo.toml
@@ -10,12 +10,18 @@ rust-version = "1.77.2"
 exclude = ["/examples", "/dist-js", "/guest-js", "/node_modules"]
 links = "tauri-plugin-websearch"
 
+[features]
+default = ["tauri"]
+# The Tauri command/plugin surface. Disabled by the headless `jan` CLI, which
+# consumes `provider` directly and must not link any GUI crate.
+tauri = ["dep:tauri"]
+
 [dependencies]
 async-trait = "0.1"
 log = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-tauri = { version = "2.5.0", default-features = false, features = [] }
+tauri = { version = "2.5.0", default-features = false, features = [], optional = true }
 thiserror = "2.0.12"
 tokio = { version = "1", features = ["full"] }
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
