+++ src-tauri/plugins/tauri-plugin-websearch/src/lib.rs
@@ -1,17 +1,15 @@
-use tauri::{
-    plugin::{Builder, TauriPlugin},
-    Runtime,
-};
-
+#[cfg(feature = "tauri")]
 mod commands;
 pub mod provider;
 
+#[cfg(feature = "tauri")]
 pub use commands::{web_fetch, web_search, WebSearchError};
 pub use provider::{FetchedPage, SearchResult};
 
 /// Initializes the web search plugin.
-pub fn init<R: Runtime>() -> TauriPlugin<R> {
-    Builder::new("websearch")
+#[cfg(feature = "tauri")]
+pub fn init<R: tauri::Runtime>() -> tauri::plugin::TauriPlugin<R> {
+    tauri::plugin::Builder::new("websearch")
         .invoke_handler(tauri::generate_handler![
             commands::web_search,
             commands::web_fetch
