+++ src-tauri/src/core/agent/project.rs
@@ -209,6 +209,7 @@ pub(crate) fn set_model_in_agent_toml(path: &Path, model: &str) -> Result<(), St
 
 /// Persist `[skills].enabled` into the agent.toml at `path`, format-preserving
 /// (comments kept). An empty list clears the whitelist (= all skills enabled).
+#[cfg(any(not(feature = "cli"), test))]
 pub(crate) fn set_skills_enabled_in_agent_toml(
     path: &Path,
     enabled: &[String],
