+++ src-tauri/src/core/app/commands.rs
@@ -2,13 +2,14 @@ use std::{
     fs,
     path::{Path, PathBuf},
 };
+#[cfg(not(feature = "cli"))]
 use tauri::{AppHandle, Manager, Runtime, State};
 
-use super::{
-    constants::{CONFIGURATION_FILE_NAME, TAURI_BUNDLE_IDENTIFIER},
-    helpers::copy_dir_recursive,
-    models::AppConfiguration,
-};
+use super::constants::{CONFIGURATION_FILE_NAME, TAURI_BUNDLE_IDENTIFIER};
+use super::models::AppConfiguration;
+#[cfg(not(feature = "cli"))]
+use super::helpers::copy_dir_recursive;
+#[cfg(not(feature = "cli"))]
 use crate::core::state::AppState;
 
 /// Canonical Jan app support directory (`%APPDATA%/Jan` on Windows).
@@ -78,6 +79,7 @@ fn legacy_app_config_candidate_paths(app_data_dir: &Path) -> Vec<PathBuf> {
     paths
 }
 
+#[cfg(not(feature = "cli"))]
 fn app_data_dir_with_fallback<R: Runtime>(app_handle: &tauri::AppHandle<R>) -> PathBuf {
     let package_name = env!("CARGO_PKG_NAME");
     app_handle
@@ -150,6 +152,7 @@ pub fn resolve_jan_data_folder() -> PathBuf {
     PathBuf::from(home).join(&app_name).join("data")
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn get_app_configurations<R: Runtime>(app_handle: tauri::AppHandle<R>) -> AppConfiguration {
     let mut app_default_configuration = AppConfiguration::default();
@@ -205,6 +208,7 @@ pub fn get_app_configurations<R: Runtime>(app_handle: tauri::AppHandle<R>) -> Ap
     }
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn update_app_configuration<R: Runtime>(
     app_handle: tauri::AppHandle<R>,
@@ -220,6 +224,7 @@ pub fn update_app_configuration<R: Runtime>(
     .map_err(|e| e.to_string())
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn get_jan_data_folder_path<R: Runtime>(app_handle: tauri::AppHandle<R>) -> PathBuf {
     if cfg!(test) {
@@ -250,6 +255,7 @@ pub fn get_jan_data_folder_path<R: Runtime>(app_handle: tauri::AppHandle<R>) -> 
     PathBuf::from(app_configurations.data_folder)
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn get_configuration_file_path<R: Runtime>(app_handle: tauri::AppHandle<R>) -> PathBuf {
     let app_path = app_data_dir_with_fallback(&app_handle);
@@ -259,6 +265,7 @@ pub fn get_configuration_file_path<R: Runtime>(app_handle: tauri::AppHandle<R>) 
     app_path.join(CONFIGURATION_FILE_NAME)
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn default_data_folder_path<R: Runtime>(app_handle: tauri::AppHandle<R>) -> String {
     let mut path = app_handle.path().data_dir().unwrap_or_else(|err| {
@@ -286,11 +293,13 @@ pub fn default_data_folder_path<R: Runtime>(app_handle: tauri::AppHandle<R>) -> 
     path_str
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn get_user_home_path<R: Runtime>(app: AppHandle<R>) -> String {
     get_app_configurations(app.clone()).data_folder
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn change_app_data_folder<R: Runtime>(
     app_handle: tauri::AppHandle<R>,
@@ -334,6 +343,7 @@ pub fn change_app_data_folder<R: Runtime>(
     update_app_configuration(app_handle, configuration)
 }
 
+#[cfg(not(feature = "cli"))]
 #[tauri::command]
 pub fn app_token(state: State<'_, AppState>) -> Option<String> {
     state.app_token.clone()
