+++ src-tauri/src/core/app/mod.rs
@@ -1,5 +1,8 @@
 pub mod commands;
 pub mod constants;
+#[cfg(not(feature = "cli"))]
 pub mod helpers;
 pub mod models;
+// Desktop settings.json store, driven by Tauri commands + a flush thread.
+#[cfg(not(feature = "cli"))]
 pub mod settings_store;
