--- src-tauri/src/core/cli/preset.rs
@@ -1,216 +1,0 @@
-//! CLI port of the desktop llamacpp `router.preset.ini` generator.
-//!
-//! The desktop app builds this file in `extensions/llamacpp-extension/src/preset.ts`
-//! from its live provider config plus every model's `model.yml`. The CLI has no
-//! live provider store, so this port emits a focused, always-loadable subset:
-//! per-model sections resolved from each `model.yml` (compiled llama.cpp defaults
-//! apply for the engine-wide `[*]` section). Global tuning knobs, MTP speculative
-//! decoding, and sampling defaults are intentionally omitted here (they require
-//! the provider config / a backend-build probe the CLI doesn't have); a preset
-//! generated by the desktop app is always preferred and never overwritten.
-
-use std::path::PathBuf;
-
-use serde_yaml::Value;
-
-use crate::core::app::commands::resolve_jan_data_folder;
-
-const DEFAULT_EMBEDDING_UBATCH: i64 = 2048;
-
-/// Ensure `<jan_data>/llamacpp/router.preset.ini` exists, generating it from the
-/// installed models when absent. Never overwrites an existing preset (the
-/// desktop-generated one carries richer engine-wide tuning). Returns its path.
-pub fn ensure_router_preset() -> Result<PathBuf, String> {
-    let data_folder = resolve_jan_data_folder();
-    let preset_path = data_folder.join("llamacpp").join("router.preset.ini");
-    if preset_path.exists() {
-        return Ok(preset_path);
-    }
-
-    let content = generate_preset()?;
-    let parent = preset_path
-        .parent()
-        .ok_or_else(|| "invalid preset path".to_string())?;
-    std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
-
-    // Atomic write (tmp + rename) so a partial file is never read by the router.
-    let tmp = preset_path.with_extension("ini.tmp");
-    std::fs::write(&tmp, content).map_err(|e| e.to_string())?;
-    std::fs::rename(&tmp, &preset_path).map_err(|e| e.to_string())?;
-    Ok(preset_path)
-}
-
-/// Build the full preset body from every discovered llamacpp model.
-fn generate_preset() -> Result<String, String> {
-    let data_folder = resolve_jan_data_folder();
-    let models_root = data_folder.join("llamacpp").join("models");
-
-    let mut sections: Vec<String> = Vec::new();
-    for (model_id, _) in super::list_models("llamacpp") {
-        let yml_path = models_root.join(&model_id).join("model.yml");
-        let raw = match std::fs::read_to_string(&yml_path) {
-            Ok(r) => r,
-            Err(_) => continue,
-        };
-        let yml: Value = match serde_yaml::from_str(&raw) {
-            Ok(v) => v,
-            Err(_) => continue,
-        };
-        if let Some(section) = model_section(&model_id, &yml) {
-            sections.push(section);
-        }
-    }
-
-    if sections.is_empty() {
-        return Err(format!(
-            "no llamacpp models found under {}; download a model with `jan pull <repo>` first",
-            models_root.display()
-        ));
-    }
-
-    Ok(sections.join("\n"))
-}
-
-/// Render one `[<model_id>]` section, or `None` when the yml has no `model_path`.
-fn model_section(model_id: &str, yml: &Value) -> Option<String> {
-    let model_path = str_field(yml, "model_path").filter(|s| !s.is_empty())?;
-
-    let mut lines = vec![format!("[{model_id}]")];
-    lines.push(format!("model = {}", resolve_path(&model_path)));
-
-    if let Some(mmproj) = str_field(yml, "mmproj_path").filter(|s| !s.is_empty()) {
-        lines.push(format!("mmproj = {}", resolve_path(&mmproj)));
-    }
-    if let Some(tpl) = str_field(yml, "chat_template").filter(|s| !s.is_empty()) {
-        lines.push(format!("chat-template = {}", escape_ini_value(&tpl)));
-    }
-    if let Some(ngl) = int_field(yml, "n_gpu_layers").filter(|&n| n >= 0) {
-        lines.push(format!("n-gpu-layers = {ngl}"));
-    }
-
-    if bool_field(yml, "embedding") {
-        lines.push("embeddings = true".to_string());
-        lines.push("ctx-size = 0".to_string());
-        let pooling = str_field(yml, "pooling")
-            .filter(|s| !s.is_empty())
-            .unwrap_or_else(|| "mean".to_string());
-        lines.push(format!("pooling = {}", escape_ini_value(&pooling)));
-        let ubatch = int_field(yml, "ubatch_size")
-            .filter(|&n| n > 0)
-            .unwrap_or(DEFAULT_EMBEDDING_UBATCH);
-        let batch = int_field(yml, "batch_size").unwrap_or(0).max(ubatch);
-        lines.push(format!("ubatch-size = {ubatch}"));
-        lines.push(format!("batch-size = {batch}"));
-    }
-
-    lines.push("load-on-startup = false".to_string());
-    lines.push(String::new());
-    Some(lines.join("\n"))
-}
-
-/// Resolve a `model.yml` path: absolute used verbatim, relative joined with the
-/// Jan data folder (mirrors `resolve_model_by_id`).
-fn resolve_path(p: &str) -> String {
-    let pb = PathBuf::from(p);
-    if pb.is_absolute() {
-        pb.to_string_lossy().into_owned()
-    } else {
-        resolve_jan_data_folder()
-            .join(p)
-            .to_string_lossy()
-            .into_owned()
-    }
-}
-
-/// Collapse newlines to a single space and trim (mirrors `escapeIniValue`).
-fn escape_ini_value(v: &str) -> String {
-    let mut out = String::with_capacity(v.len());
-    let mut prev_space = false;
-    for ch in v.chars() {
-        if ch == '\r' || ch == '\n' {
-            if !prev_space {
-                out.push(' ');
-                prev_space = true;
-            }
-        } else {
-            out.push(ch);
-            prev_space = false;
-        }
-    }
-    out.trim().to_string()
-}
-
-fn str_field(yml: &Value, key: &str) -> Option<String> {
-    yml.get(key).and_then(|v| v.as_str()).map(str::to_owned)
-}
-
-fn int_field(yml: &Value, key: &str) -> Option<i64> {
-    yml.get(key).and_then(|v| v.as_i64())
-}
-
-fn bool_field(yml: &Value, key: &str) -> bool {
-    yml.get(key).and_then(|v| v.as_bool()).unwrap_or(false)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn yml(s: &str) -> Value {
-        serde_yaml::from_str(s).unwrap()
-    }
-
-    #[test]
-    fn chat_model_section_has_model_and_startup() {
-        let v = yml("model_path: /abs/model.gguf\nname: Foo\n");
-        let section = model_section("foo", &v).unwrap();
-        assert!(section.starts_with("[foo]\n"));
-        assert!(section.contains("model = /abs/model.gguf"));
-        assert!(section.contains("load-on-startup = false"));
-        assert!(!section.contains("embeddings"));
-    }
-
-    #[test]
-    fn absolute_model_path_used_verbatim() {
-        assert_eq!(resolve_path("/abs/x.gguf"), "/abs/x.gguf");
-    }
-
-    #[test]
-    fn relative_model_path_joined_with_data_folder() {
-        let resolved = resolve_path("llamacpp/models/foo/model.gguf");
-        assert!(resolved.ends_with("llamacpp/models/foo/model.gguf"));
-        assert!(PathBuf::from(&resolved).is_absolute());
-    }
-
-    #[test]
-    fn vision_model_emits_mmproj() {
-        let v = yml("model_path: /abs/m.gguf\nmmproj_path: /abs/mmproj.gguf\n");
-        let section = model_section("v", &v).unwrap();
-        assert!(section.contains("mmproj = /abs/mmproj.gguf"));
-    }
-
-    #[test]
-    fn embedding_model_emits_embedding_block() {
-        let v = yml(
-            "model_path: /abs/e.gguf\nembedding: true\npooling: mean\nubatch_size: 2048\nbatch_size: 1024\n",
-        );
-        let section = model_section("e", &v).unwrap();
-        assert!(section.contains("embeddings = true"));
-        assert!(section.contains("ctx-size = 0"));
-        assert!(section.contains("pooling = mean"));
-        assert!(section.contains("ubatch-size = 2048"));
-        // batch-size is floored to ubatch-size.
-        assert!(section.contains("batch-size = 2048"));
-    }
-
-    #[test]
-    fn missing_model_path_yields_no_section() {
-        let v = yml("name: Foo\n");
-        assert!(model_section("foo", &v).is_none());
-    }
-
-    #[test]
-    fn escape_collapses_newlines() {
-        assert_eq!(escape_ini_value("a\n b \r\n c "), "a  b   c");
-    }
-}
