+++ src-tauri/src/core/mcp/tests.rs
@@ -763,116 +763,6 @@ fn test_server_summary_round_trip() {
 }
 
 // ============================================================================
-// helpers.rs pure-function tests: extract_command_args / extract_active_status
-// ============================================================================
-
-#[test]
-fn test_extract_command_args_minimal_config() {
-    use super::helpers::extract_command_args;
-    let cfg = serde_json::json!({
-        "command": "npx",
-        "args": ["-y", "server"]
-    });
-    let parsed = extract_command_args(&cfg).expect("should parse");
-    assert_eq!(parsed.command, "npx");
-    assert_eq!(parsed.args.len(), 2);
-    assert_eq!(parsed.args[0], "-y");
-    assert!(parsed.url.is_none());
-    assert!(parsed.transport_type.is_none());
-    assert!(parsed.timeout.is_none());
-    assert!(parsed.envs.is_empty());
-    assert!(parsed.headers.is_empty());
-}
-
-#[test]
-fn test_extract_command_args_full_config() {
-    use super::helpers::extract_command_args;
-    let cfg = serde_json::json!({
-        "command": "",
-        "args": [],
-        "type": "http",
-        "url": "https://mcp.example.com/mcp",
-        "timeout": 45,
-        "env": {"API_KEY": "abc", "DEBUG": "1"},
-        "headers": {"Authorization": "Bearer xyz"}
-    });
-    let parsed = extract_command_args(&cfg).expect("should parse");
-    assert_eq!(parsed.command, "");
-    assert_eq!(parsed.transport_type.as_deref(), Some("http"));
-    assert_eq!(parsed.url.as_deref(), Some("https://mcp.example.com/mcp"));
-    assert_eq!(parsed.timeout, Some(Duration::from_secs(45)));
-    assert_eq!(
-        parsed.envs.get("API_KEY").and_then(|v| v.as_str()),
-        Some("abc")
-    );
-    assert_eq!(parsed.envs.get("DEBUG").and_then(|v| v.as_str()), Some("1"));
-    assert_eq!(
-        parsed.headers.get("Authorization").and_then(|v| v.as_str()),
-        Some("Bearer xyz")
-    );
-}
-
-#[test]
-fn test_extract_command_args_returns_none_when_required_fields_missing() {
-    use super::helpers::extract_command_args;
-    // Missing command
-    let cfg = serde_json::json!({"args": []});
-    assert!(extract_command_args(&cfg).is_none());
-    // Missing args
-    let cfg = serde_json::json!({"command": "npx"});
-    assert!(extract_command_args(&cfg).is_none());
-    // Not an object
-    let cfg = serde_json::json!(["a", "b"]);
-    assert!(extract_command_args(&cfg).is_none());
-    // command not a string
-    let cfg = serde_json::json!({"command": 123, "args": []});
-    assert!(extract_command_args(&cfg).is_none());
-    // args not an array
-    let cfg = serde_json::json!({"command": "npx", "args": "oops"});
-    assert!(extract_command_args(&cfg).is_none());
-}
-
-#[test]
-fn test_extract_command_args_parses_default_mcp_config_servers() {
-    use super::constants::DEFAULT_MCP_CONFIG;
-    use super::helpers::extract_command_args;
-    let value: serde_json::Value = serde_json::from_str(DEFAULT_MCP_CONFIG).unwrap();
-    for (name, cfg) in value["mcpServers"].as_object().unwrap() {
-        let parsed = extract_command_args(cfg)
-            .unwrap_or_else(|| panic!("default config server '{name}' should parse"));
-        // command may be empty for HTTP transports
-        if name == "exa" {
-            assert_eq!(parsed.transport_type.as_deref(), Some("http"));
-            assert!(parsed.url.is_some());
-        } else {
-            assert!(!parsed.command.is_empty(), "{name} should have a command");
-        }
-    }
-}
-
-#[test]
-fn test_extract_active_status_variants() {
-    use super::helpers::extract_active_status;
-    assert_eq!(
-        extract_active_status(&serde_json::json!({"active": true})),
-        Some(true)
-    );
-    assert_eq!(
-        extract_active_status(&serde_json::json!({"active": false})),
-        Some(false)
-    );
-    // Missing
-    assert_eq!(extract_active_status(&serde_json::json!({})), None);
-    // Wrong type
-    assert_eq!(
-        extract_active_status(&serde_json::json!({"active": "yes"})),
-        None
-    );
-    // Not an object
-    assert_eq!(extract_active_status(&serde_json::json!(true)), None);
-}
-
-// ============================================================================
 // lockfile.rs Tests
 // ============================================================================
