+++ src-tauri/src/core/state.rs
@@ -1,20 +1,23 @@
-use std::{
-    collections::{HashMap, HashSet},
-    sync::Arc,
-};
+use std::collections::HashMap;
+#[cfg(not(feature = "cli"))]
+use std::collections::HashSet;
+use std::sync::Arc;
 
-use crate::core::{
-    downloads::models::DownloadManagerState,
-    mcp::models::{McpSettings, ToolWithServer},
-};
+#[cfg(not(feature = "cli"))]
+use crate::core::downloads::models::DownloadManagerState;
+#[cfg(not(feature = "cli"))]
+use crate::core::mcp::models::{McpSettings, ToolWithServer};
 use rmcp::{
     model::{CallToolRequestParam, CallToolResult, InitializeRequestParam, Tool},
     service::RunningService,
     RoleClient, ServiceError,
 };
-use tokio::sync::{oneshot, Mutex, Notify};
+use tokio::sync::Mutex;
+#[cfg(not(feature = "cli"))]
+use tokio::sync::{oneshot, Notify};
 
 /// Server handle type for managing the proxy server lifecycle
+#[cfg(not(feature = "cli"))]
 pub type ServerHandle =
     tokio::task::JoinHandle<Result<(), Box<dyn std::error::Error + Send + Sync>>>;
 
@@ -60,6 +63,9 @@ pub enum RunningServiceEnum {
 }
 pub type SharedMcpServers = Arc<Mutex<HashMap<String, RunningServiceEnum>>>;
 
+/// Shared desktop application state owned by Tauri. The CLI builds its
+/// subsystems (MCP map, provider configs) directly instead.
+#[cfg(not(feature = "cli"))]
 pub struct AppState {
     pub app_token: Option<String>,
     pub mcp_servers: SharedMcpServers,
@@ -92,6 +98,7 @@ pub struct AppState {
     pub mcp_last_known_tools: Arc<Mutex<HashMap<String, Vec<ToolWithServer>>>>,
 }
 
+#[cfg(not(feature = "cli"))]
 impl Default for AppState {
     fn default() -> Self {
         Self {
